# docs/openapi/hansapp-openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/me": {
      "get": {
        "description": "현재 로그인 사용자 정보를 반환한다.",
        "operationId": "UserController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "내 정보",
        "tags": [
          "users"
        ]
      }
    },
...
    "/boards/{name}/posts": {
      "get": {
        "description": "공개된 글만. 고정 글이 먼저, 그다음 최신순.\n\n**인자 없이 부른 첫 페이지에만** `Cache-Control: public, max-age=60` 이 붙는다 — 게시판을 열면 누구나 보게 되는 한 장이라, 링크가 퍼질 때의 순간 트래픽을 CDN 이 받아 낸다. 페이지를 넘기거나 크기를 바꾸면 no-store 다. 새 글은 최대 1분 뒤에 목록에 나타난다.",
        "operationId": "BoardController_listPosts",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "페이지 번호(1부터)",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "size",
            "required": false,
            "in": "query",
            "description": "페이지 크기",
            "schema": {
              "maximum": 50,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PageResponseDto"
                    },
                    {
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/PublicPostSummaryDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "게시글 목록",
        "tags": [
          "board"
        ]
      }
    },
    "/boards/{name}/posts/{id}": {
      "get": {
        "description": "비공개 글의 본문은 쓴 사람에게만 나간다 — 나머지에게는 content 가 빠진다.\n\n**공개 글에만** `Cache-Control: public, max-age=60` 이 붙는다 — 오래 태우려는 것이 아니라 링크가 퍼질 때의 순간 트래픽을 CDN 이 받아 내기 위한 1분이다. 그동안의 조회는 서버에 닿지 않아 조회수도 그만큼 덜 는다. 비공개 글은 no-store 라 공유 캐시에 남지 않는다.",
        "operationId": "BoardController_getPost",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicPostDetailDto"
                }
              }
            }
          }
        },
        "summary": "게시글 하나",
        "tags": [
          "board"
        ]
      }
    },
...
    "/nts/business/{bno}/status": {
      "get": {
        "description": "사업자번호로 납세자상태(계속/휴업/폐업)와 과세유형을 조회한다.\n\n**등록되지 않은 번호도 에러가 아니다** — registered=false 로 응답한다.",
        "operationId": "BusinessController_status",
        "parameters": [
          {
            "name": "bno",
            "required": true,
            "in": "path",
            "description": "사업자등록번호. '-' 는 자동으로 제거된다.",
            "schema": {
              "example": "6462401820",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessStatusDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "client-id": []
          }
        ],
        "summary": "사업자 등록번호 상태조회",
        "tags": [
          "business"
        ]
      }
    },
    "/nts/business/{bno}/verify": {
      "post": {
        "description": "사업자번호(경로)·개업일자·대표자성명이 국세청 등록정보와 일치하는지 확인한다.\n개업일자·대표자성명은 필수이고, 상호·법인번호·주소를 더 넣으면 판정이 정밀해진다.\n\n대표자성명 등 개인정보는 URL·쿼리가 아니라 **본문**으로 받는다(접근 로그 노출 방지).",
        "operationId": "BusinessController_verify",
        "parameters": [
          {
            "name": "bno",
            "required": true,
            "in": "path",
            "description": "사업자등록번호. '-' 는 자동으로 제거된다.",
            "schema": {
              "example": "6462401820",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BusinessVerifyRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessVerificationDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "client-id": []
          }
        ],
        "summary": "사업자 등록정보 진위확인",
        "tags": [
          "business"
        ]
      }
    },
...
  },
  "info": {
    "title": "Hans API",
    "description": "Hans API backend 문서",
    "version": "0.15.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://127.0.0.1:3000",
      "description": "Local"
    },
    {
      "url": "https://develop-api.plzhans.com",
      "description": "Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT or service key(sk_...)",
        "type": "http",
        "description": "사용자 access token(JWT) 또는 서비스 키(sk_{appId}_{keyId}_{rand})."
      },
      "client-id": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Client-Id",
        "description": "공개 클라이언트 식별자(clientId). WEB 은 Origin 도 검사됨."
      }
    },
    "schemas": {
...
      "MeResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "회원번호"
          },
          "email": {
            "type": "string",
            "description": "이메일"
          },
          "emailVerified": {
            "type": "boolean",
            "description": "이메일 인증 여부"
          },
          "name": {
            "type": "object",
            "description": "표시 이름"
          },
          "role": {
            "type": "string",
            "description": "권한",
            "example": "USER"
          },
          "tier": {
            "type": "string",
            "description": "회원 등급. 앱 생성 한도를 정한다.",
            "enum": [
              "BASIC",
              "PRO",
              "UNLIMITED"
            ],
            "example": "BASIC"
          },
          "joinType": {
            "type": "string",
            "description": "최초 가입 수단",
            "example": "EMAIL"
          },
          "language": {
            "type": "string",
            "description": "표시·메일 언어. 고른 적이 없으면 null 이고, 그때는 요청의 Accept-Language 를 따른다.",
            "enum": [
              "ko",
              "en",
              "ja",
              "zh"
            ],
            "example": "ko"
          },
          "timeZone": {
            "type": "string",
            "description": "IANA 타임존 ID. 고른 적이 없으면 null 이다.",
            "example": "Asia/Seoul"
          },
          "createdAt": {
            "type": "string",
            "description": "가입 일시",
            "example": "2026-08-06T02:10:00.000Z"
          },
          "hasPassword": {
            "type": "boolean",
            "description": "비밀번호가 설정돼 있는가. 소셜로만 가입한 계정은 false 라 화면이 \"변경\" 대신 \"설정\" 을 띄운다."
          },
          "linkedProviders": {
            "description": "연동된 소셜 제공자 목록",
            "example": [
              "GOOGLE",
              "KAKAO"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "email",
          "emailVerified",
          "role",
          "tier",
          "joinType",
          "createdAt",
          "hasPassword",
          "linkedProviders"
        ]
      },
...
      "PageResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "description": "목록 항목",
            "type": "array",
            "items": {
              "type": "array"
            }
          },
          "page": {
            "type": "number",
            "description": "현재 페이지 번호"
          },
          "size": {
            "type": "number",
            "description": "페이지 크기"
          },
          "totalCount": {
            "type": "number",
            "description": "전체 항목 수"
          },
          "totalPages": {
            "type": "number",
            "description": "전체 페이지 수"
          }
        },
        "required": [
          "items",
          "page",
          "size",
          "totalCount",
          "totalPages"
        ]
      },
      "PublicAuthorDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "number",
            "enum": [
              1,
              2
            ]
          },
          "id": {
            "type": "number",
            "description": "회원 글이면 회원번호, 운영자 글이면 0"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "id",
          "name"
        ]
      },
      "PublicPostSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "title": {
            "type": "string"
          },
          "summary": {
            "type": "object",
            "nullable": true
          },
          "author": {
            "$ref": "#/components/schemas/PublicAuthorDto"
          },
          "pinned": {
            "type": "boolean"
          },
          "secret": {
            "type": "boolean",
            "description": "비공개 글인가"
          },
          "commentEnabled": {
            "type": "boolean",
            "description": "게시판·글 설정을 합친 결과"
          },
          "likeEnabled": {
            "type": "boolean",
            "description": "게시판·글 설정을 합친 결과"
          },
          "publishedAt": {
            "type": "object",
            "nullable": true
          },
          "viewCount": {
            "type": "number"
          },
          "commentCount": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "title",
          "summary",
          "author",
          "pinned",
          "secret",
          "commentEnabled",
          "likeEnabled",
          "publishedAt",
          "viewCount",
          "commentCount"
        ]
      },
      "PublicCommentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "parentId": {
            "type": "object",
            "nullable": true,
            "description": "답글이면 부모 댓글 번호"
          },
          "author": {
            "$ref": "#/components/schemas/PublicAuthorDto"
          },
          "content": {
            "type": "object",
            "nullable": true
          },
          "secret": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "parentId",
          "author",
          "secret",
          "createdAt"
        ]
      },
      "PublicPostDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "title": {
            "type": "string"
          },
          "summary": {
            "type": "object",
            "nullable": true
          },
          "author": {
            "$ref": "#/components/schemas/PublicAuthorDto"
          },
          "pinned": {
            "type": "boolean"
          },
          "secret": {
            "type": "boolean",
            "description": "비공개 글인가"
          },
          "commentEnabled": {
            "type": "boolean",
            "description": "게시판·글 설정을 합친 결과"
          },
          "likeEnabled": {
            "type": "boolean",
            "description": "게시판·글 설정을 합친 결과"
          },
          "publishedAt": {
            "type": "object",
            "nullable": true
          },
          "viewCount": {
            "type": "number"
          },
          "commentCount": {
            "type": "number"
          },
          "content": {
            "type": "object",
            "description": "본문(마크다운)"
          },
          "comments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicCommentDto"
            }
          }
        },
        "required": [
          "id",
          "title",
          "summary",
          "author",
          "pinned",
          "secret",
          "commentEnabled",
          "likeEnabled",
          "publishedAt",
          "viewCount",
          "commentCount",
          "comments"
        ]
      },
...
      "BusinessStatusDto": {
        "type": "object",
        "properties": {
          "bno": {
            "type": "string",
            "description": "사업자등록번호",
            "example": "6462401820"
          },
          "registered": {
            "type": "boolean",
            "description": "국세청에 등록된 사업자인지. false 면 나머지 상태값은 비어 있다.",
            "example": true
          },
          "statusCode": {
            "type": "string",
            "description": "납세자상태 코드. 01:계속사업자 / 02:휴업자 / 03:폐업자",
            "example": "01"
          },
          "status": {
            "type": "string",
            "description": "납세자상태 명칭",
            "example": "계속사업자"
          },
          "taxTypeCode": {
            "type": "string",
            "description": "과세유형 코드. 01:일반 / 02:간이 / 04:면세 등",
            "example": "01"
          },
          "taxType": {
            "type": "string",
            "description": "과세유형 명칭. 미등록이면 \"국세청에 등록되지 않은 사업자등록번호입니다\".",
            "example": "부가가치세 일반과세자"
          },
          "closedAt": {
            "type": "string",
            "description": "폐업일 (YYYYMMDD)",
            "example": "20200101"
          }
        },
        "required": [
          "bno",
          "registered"
        ]
      },
      "BusinessVerifyRequestDto": {
        "type": "object",
        "properties": {
          "startDate": {
            "type": "string",
            "description": "개업일자 (YYYYMMDD). '-' 는 자동 제거된다.",
            "example": "20000101"
          },
          "name": {
            "type": "string",
            "description": "대표자성명",
            "example": "홍길동"
          },
          "name2": {
            "type": "string",
            "description": "대표자성명2. 대표자성명이 한글이 아닐 때의 한글명."
          },
          "companyName": {
            "type": "string",
            "description": "상호"
          },
          "corpNo": {
            "type": "string",
            "description": "법인등록번호(13자리). '-' 는 자동 제거."
          },
          "address": {
            "type": "string",
            "description": "사업장주소"
          }
        },
        "required": [
          "startDate",
          "name"
        ]
      },
      "BusinessVerificationDto": {
        "type": "object",
        "properties": {
          "bno": {
            "type": "string",
            "description": "사업자등록번호",
            "example": "6462401820"
          },
          "valid": {
            "type": "boolean",
            "description": "진위확인 결과. 입력한 정보가 국세청 등록정보와 일치하면 true.",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "불일치 시 안내 메시지. 일치하면 비어 있다.",
            "example": "확인할 수 없습니다."
          },
          "status": {
            "description": "함께 조회된 사업자 상태.",
            "allOf": [
              {
                "$ref": "#/components/schemas/BusinessStatusDto"
              }
            ]
          }
        },
        "required": [
          "bno",
          "valid"
        ]
      },
...
    }
  }
}